requirements.txt
torch>=2.2
transformers>=4.40
datasets>=2.18
peft>=0.10
trl>=0.8
accelerate>=0.29
pyyaml>=6.0
pydantic>=2.6
fastapi>=0.110
uvicorn>=0.29
sentencepiece>=0.2
protobuf>=5.26